requirements.txt
fastapi
uvicorn[standard]
groq
requests
pydantic
playwright
pymongo
python-dotenv
python-multipart